Make local vars visible in nested scopes
The behaviour of `eval`/`exec` in Python 3 is totally weird:
https://bugs.python.org/issue37646 Variables passed in `locals` are
visible to top-level scopes, but invisible in nested scopes
(function/class definitions, comprehensions). To allow, e.g., functions
which can update or check global buffers, we use this egregious hack:
pass *all* variables in as globals. However, *updates* to variables in
`exec` calls are applied to the `locals` mapping, so we must retain
`self.locals`.